+++ Cadmus.Core/Config/ItemMetadataBuilderFactory.cs
@@ -1,0 +1,130 @@
+﻿using Fusi.Tools.Configuration;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+
+namespace Cadmus.Core.Config;
+
+/// <summary>
+/// A factory for <see cref="IItemMetadataBuilder"/>'s. This factory relies on
+/// a configuration rooted at the <c>metadataBuilders</c> section, including an
+/// optional <c>options</c> subsection for each builder.
+/// </summary>
+public sealed class ItemMetadataBuilderFactory : ComponentFactory
+{
+    private readonly string _connectionString;
+
+    /// <summary>
+    /// The name of the connection string property to be supplied
+    /// in POCO option objects (<c>ConnectionString</c>).
+    /// </summary>
+    public const string CONNECTION_STRING_NAME = "ConnectionString";
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ItemMetadataBuilderFactory"/>
+    /// class.
+    /// </summary>
+    /// <param name="host">The host.</param>
+    /// <param name="connectionString">The general connection string to
+    /// supply to any component requiring an option named
+    /// <see cref="CONNECTION_STRING_NAME"/> (=<c>ConnectionString</c>),
+    /// when this option is not specified in its configuration.</param>
+    public ItemMetadataBuilderFactory(IHost host, string connectionString) :
+        base(host)
+    {
+        _connectionString = connectionString ??
+            throw new ArgumentNullException(nameof(connectionString));
+    }
+
+    /// <summary>
+    /// Overrides the options.
+    /// </summary>
+    /// <param name="options">The options.</param>
+    /// <param name="section">The section.</param>
+    protected override void OverrideOptions(object options,
+        IConfigurationSection? section)
+    {
+        Type optionType = options.GetType();
+
+        // if we have a default connection AND the options type
+        // has a ConnectionString property, see if we should supply a value
+        // for it
+        PropertyInfo? property;
+        if (_connectionString != null &&
+            (property = optionType.GetProperty(CONNECTION_STRING_NAME)) != null)
+        {
+            // here we can safely discard the returned object as it will
+            // be equal to the input options, which is not null
+            SupplyProperty(optionType, property, options, _connectionString);
+        }
+    }
+
+    /// <summary>
+    /// Configures the container services to use components from
+    /// <c>Cadmus.Core</c> and <c>Cadmus.Seed</c>, plus the assemblies
+    /// specified by <paramref name="additionalAssemblies"/>.
+    /// This is just a helper method: at any rate, the configuration of
+    /// the container is external to the VSM factory. You could use this
+    /// method as a model and create your own, or call this method to
+    /// register the components from these two assemblies, and then
+    /// further configure the container, or add more assemblies when
+    /// calling this via <paramref name="additionalAssemblies"/>.
+    /// </summary>
+    /// <param name="services">The services.</param>
+    /// <param name="partTypeProvider">The part type provider.</param>
+    /// <param name="additionalAssemblies">The optional additional
+    /// assemblies.</param>
+    /// <exception cref="ArgumentNullException">container or part type
+    /// provider</exception>
+    public static void ConfigureServices(IServiceCollection services,
+        IPartTypeProvider partTypeProvider,
+        params Assembly[] additionalAssemblies)
+    {
+        ArgumentNullException.ThrowIfNull(services);
+        ArgumentNullException.ThrowIfNull(partTypeProvider);
+
+        // https://simpleinjector.readthedocs.io/en/latest/advanced.html?highlight=batch#batch-registration
+        Assembly[] assemblies = additionalAssemblies ?? [];
+
+        services.AddSingleton(partTypeProvider);
+
+        foreach (Type it in new[]
+        {
+            typeof(IItemMetadataBuilder),
+        })
+        {
+            foreach (Type t in GetAssemblyConcreteTypes(assemblies, it))
+            {
+                services.AddTransient(it, t);
+            }
+        }
+    }
+
+    /// <summary>
+    /// Gets from <c>metadataBuilders</c> the metadata builder matching
+    /// the specified facet ID. In this section, each configurable object
+    /// representing a builder has its Keys property set to a list of strings,
+    /// each one being either a facet ID.
+    /// </summary>
+    /// <returns>Item metadata builder or null if no match is found.</returns>
+    public IItemMetadataBuilder? GetItemMetadataBuilder(string facetId)
+    {
+        IList<ComponentFactoryConfigEntry> entries =
+            ComponentFactoryConfigEntry.ReadComponentEntries(
+            Configuration, "metadataBuilders");
+
+        ComponentFactoryConfigEntry? entry =
+            entries.FirstOrDefault(e => e.Keys?.Contains(facetId) == true);
+        if (entry == null) return null;
+
+        int i = entries.IndexOf(entry);
+        return GetComponent<IItemMetadataBuilder>(
+            facetId,
+            $"metadataBuilders:{i}:options",
+            false);
+    }
+}